apps/api/pyproject.toml
[project]
name = "stepsquad-api"
version = "0.4.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "python-dateutil>=2.9.0.post0",
  "google-cloud-firestore>=2.16.0",
  "google-cloud-bigquery>=3.25.0",
  "google-cloud-pubsub>=2.21.0",
]
